Introduce new skill system subsystem stuff
* Adds a new repository pattern, new skill packets and adds new
possibilities for new skill types

+++ OpenORPG.Database/Models/ContentTemplates/ItemTemplate.cs
@@ -61,7 +61,7 @@ namespace Server.Game.Database.Models.ContentTemplates
         /// <summary>
         /// A skill ID that a user might learn by using this item.
         /// </summary>
-        public long LearntSkillId { get; set; }
+        public int LearntSkillId { get; set; }
      
 
     }
